provide a setup script and explain installation in README
The advantage of a setup script is backward compatibility. We might
want to do more initialisation in there in the future (e.g. provide
a flag that all functions should be readonly).
If we move this into the setup.sh then we can change it in the future
without the consumers needing to change their source code.