parse-args print help only on demand to clutter less the output
I realised that I am used to scroll up in case of errors to see the
first message, i.e. the idea to print the help first and the error in
the end was rather counterintuitive.
Let's take a new approach and print the error first again and ask the
user if the help shall be printed, this way we only show it on demand
and don't clutter the output -- i.e. still see the error first without
losing the ability to see the help quickly