gpg < 2.4 limits path to 180 chars, symlink gpg dir during import
The actual error occurs when the gpg-agent tries to create a unix socket
Per default a gpg-agent is not started for things like --verify, --list
etc. only when needing access to a private key (such as importing one).
Hence, we apply the workaround only for importGpgKey and not in other
cases where --homedir is used

This also means we need to rewrite trustGpgKey as it currently uses
--edit-key which again needs access to the private key => use
--import-ownertrust instead, which again should not use a gpg-agent.